refactor(11-03): extract document endpoints into documents_router to break psycopg2 import chain
Moves GET /tool-arena/documents and GET /tool-arena/documents/{doc_id} into
a dedicated sub-router (documents_router.py) that only imports documents.py.
router.py delegates via include_router. test_documents_endpoints.py now imports
documents_router directly — no psycopg2/Redis deps in the test collection path.

16/16 document tests pass (9 registry + 7 endpoint).
